__mg_time -> mg_os_time; __mg_mktime -> __mg_os_mktime; __mg_localtime -> __mg_os_localtime; use clock_gettime for tick count instead of gettimeofday

+++ include/common.h
@@ -1991,7 +1991,7 @@ extern "C" {
 #endif
 
 #ifndef HAVE_TIME
-typedef unsigned int time_t;
+typedef unsigned long time_t;
 
 struct tm {
     int    tm_sec;   /* seconds [0,61] */
